pyproject.toml
[project]
name = "inverter-api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Joël Kuijper", email = "jmkuijper@outlook.com" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "python-multipart>=0.0.9",
    "python-pptx>=0.6.23",
    "pillow>=10.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye.scripts]
dev = "uvicorn src.main:app --reload"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]
